Prevent alter db from failing if db doesn't have location
When creating DB from glue UI (and potentially some other means),
the Db location is not mandatory to set, and is also not mocked
(okera does mock it if you don't specify it).

This exposes a weird bug in Amazon's API where alter db calls fail
if your db does not have a location leading to a catch-22 where you
cannot set a location, because the alter api won't work. In order to avoid
this bug failing calls to alter database, if your database object does
not have a location, set it as part of your `alter database` call