[DAS-5082] Make doesDefaultDBExist call the database existence checker in  common
*What this PR does / why we need it:*
The spark jar code randomly calls glue RPCs, when in fact the correct flow should be to
utilize the existing helpers in the common jar or implement them if needed as much as possible.

This became imperative as a blocker for the "cache okera system tables and dbs" effort, as without
this, we'd need to cache in the common and spark layer, leading to a variety of validity / cache coherence
issues.

To do this, fix the databaseExists function in common, and use it from the spark jar.

*Special notes for your reviewer:*

*Additional testing that was done:*

```
Release note:
```