[DAS-5181] Forbid creation and detection of existence of default.okera_system
For some reason this table gets looked up often. While this may appear a hacky fix, I
feel a sustainable way to save some costly RPCs is to simply return a false when an
existence check for default.okera_system happens.

This does leave the question of what if someone tries to create this table: we address this
by explicitly forbidding and retuning an error when someone attempts to create a table in
the default db by the name okera_system.

Please note that the default db in this case does not have to be "default". It can be whatever
the HMS version code beeing picked up, uses as the default db, which currently happens to be
"default"